requirements.txt
llama-cpp-python==0.1.49
